update readme update availability controller added body and parameter check for missing parameters move getEventStruct to event interface move getUserStruct to user interface update eventRoute to remove getting all events update SQL queries to use capitlization for clarity remove unused imports
