+++ backend/src/interfaces/event.ts
@@ -5,3 +5,13 @@ export interface Event {
     endDate: Date;
     timezone: string;
 }
+
+export function getEventStruct(data: any): Event {
+    return {
+        eventId: data.event_id,
+        name: data.name,
+        startDate: data.start_date,
+        endDate: data.end_date,
+        timezone: data.timezone,
+    };
+}
