+++ backend/src/interfaces/user.ts
@@ -3,3 +3,7 @@ export interface User {
     eventId: number;
     name: string;
 }
+
+export function getUserStruct(data: any): User {
+    return { userId: data.user_id, eventId: data.event_id, name: data.name };
+}
