+++ backend/src/routes/eventRoute.ts
@@ -2,7 +2,7 @@ import express, { Request, Response } from "express";
 import pool from "../db";
 import { isValid } from "date-fns";
 import { EventController } from "../controllers/eventController";
-import { generateNRandomId, getUtcDateTime } from "../utils";
+import { getUtcDateTime } from "../utils";
 import { UserController } from "../controllers/userController";
 
 const router = express.Router();
@@ -10,39 +10,29 @@ const router = express.Router();
 const eventController: EventController = new EventController();
 const userController: UserController = new UserController();
 
-router
-    .route("/")
-    // .get(async (req: Request, res: Response) => {
-    //     try {
-    //         const [rows] = await pool.query("SELECT * FROM events");
-    //         res.status(200).json(rows);
-    //     } catch (error: any) {
-    //         res.status(500).json({ error: error.message });
-    //     }
-    // })
-    .post(eventController.createEvent);
+router.route("/").post(eventController.createEvent);
 
 router.route("/:eventId").get(eventController.getEvent);
 
 router.route("/:eventId/users").get(userController.getUsersFromEvent).post(userController.createUser);
 
 router
-    .route("/:eventId/users/:user_id") // drop this route? this isnt needed since events route returns the user and availability information from the request
-    .get(async (req: Request, res: Response) => {
-        try {
-            const userSql = "SELECT * FROM users where user_id = ?";
-            const availSql = "select available from availability where user_id = ?";
-            const [userRows]: any[] = await pool.query(userSql, [req.params.user_id]);
-            if (userRows.length == 0) {
-                res.status(404).json({ error: "Resource not found" });
-            }
-            const [availRows]: any[] = await pool.query(availSql, [req.params.user_id]);
-            const availability: string[] = availRows.map((val: any) => val.available);
-            res.status(200).json({ ...userRows[0], availability: availability });
-        } catch (error: any) {
-            res.status(500).json({ error: error.message });
-        }
-    })
+    .route("/:eventId/users/:userId")
+    // .get(async (req: Request, res: Response) => {
+    //     try {
+    //         const userSql = "SELECT * FROM users WHERE user_id = ?";
+    //         const availSql = "SELECT available FROM availability WHERE user_id = ?";
+    //         const [userRows]: any[] = await pool.query(userSql, [req.params.user_id]);
+    //         if (userRows.length == 0) {
+    //             res.status(404).json({ error: "Resource not found" });
+    //         }
+    //         const [availRows]: any[] = await pool.query(availSql, [req.params.user_id]);
+    //         const availability: string[] = availRows.map((val: any) => val.available);
+    //         res.status(200).json({ ...userRows[0], availability: availability });
+    //     } catch (error: any) {
+    //         res.status(500).json({ error: error.message });
+    //     }
+    // })
     .put(async (req: Request, res: Response) => {
         const times: string[] = req.body.availability;
         const timezone: string = req.body.timezone;
@@ -52,7 +42,7 @@ router
             return;
         }
         dates.forEach((date: Date) => {});
-        const sql = "insert into availability values (user_id, available) values (?, ?)";
+        const sql = "INSERT INTO availability (user_id, available) VALUES (?, ?)";
         const [rows]: any[] = await pool.query(sql, [req.params]);
     });
